+++ workflow/webpack.renderer.config.js
@@ -153,7 +153,9 @@ const rendererConfig = {
             failOnError: true,
             allowAsyncCycles: false,
             cwd: process.cwd()
-        })
+        }),
+        // Exclude moment.js's locale files.
+        new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/)
     ],
     output: {
         filename: '[name].js',