+++ support/release-notes.md
@@ -2,5 +2,7 @@ New:
     - Progress indicator in application icon
     - More controls for running frameworks in bulk, including from repository or project context menus in the sidebar
 Changed:
-    - Pre-compile main process for faster loading times - #14
+    - Better formatting of duration strings
+    - Less dependency on Node APIs in the renderer process
     - Updated Primer CSS and Octicons to latest version
+    - Pre-compile main process for faster loading times - #14